Fixed parser errors
