+++ src/Core/Language/Utf8/Utf8GraphQLRequestParser.Request.cs
@@ -1,0 +1,21 @@
+using System;
+using System.Collections.Generic;
+
+namespace HotChocolate.Language
+{
+    public ref partial struct Utf8GraphQLRequestParser
+    {
+        private ref struct Request
+        {
+            public string OperationName { get; set; }
+
+            public string NamedQuery { get; set; }
+
+            public ReadOnlySpan<byte> Query { get; set; }
+
+            public IReadOnlyDictionary<string, object> Variables { get; set; }
+
+            public IReadOnlyDictionary<string, object> Extensions { get; set; }
+        }
+    }
+}
